refactored tests: alldistinct, allequal, groupby2, split, unique_justseen
