+++ tests/test_itertools_cfuncs.py
@@ -1,6 +1,5 @@
 # Built-ins
 from __future__ import absolute_import, division, print_function
-import operator
 import pickle
 
 # 3rd party
@@ -12,421 +11,6 @@ import iteration_utilities
 # Test helper
 from helper_leak import memory_leak
 from helper_pytest_monkeypatch import pytest_raises
-
-
-def test_unique_justseen():
-    unique_justseen = iteration_utilities.unique_justseen
-
-    assert list(unique_justseen([])) == []
-    assert list(unique_justseen([1, 1, 2, 3, 3])) == [1, 2, 3]
-    assert list(unique_justseen('aAabBb')) == ['a', 'A', 'a', 'b', 'B', 'b']
-    assert list(unique_justseen('aAabBb', key=str.lower)) == ['a', 'b']
-
-    with pytest.raises(TypeError):  # not iterable
-        unique_justseen(10)
-
-    with pytest.raises(TypeError):  # function call fails
-        list(unique_justseen([1, 2, 3], key=lambda x: x + 'a'))
-
-    class Test2(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            raise TypeError()
-
-        def __ne__(self, other):
-            raise TypeError()
-
-    with pytest.raises(TypeError):  # objects do not support eq or ne
-        list(unique_justseen([Test2(1), Test2(2)]))
-
-    class Test3(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            raise TypeError()
-
-        def __ne__(self, other):
-            return self.value != other.value
-
-    res = list(unique_justseen([Test3(1), Test3(1)]))
-    assert len(res) == 1
-    assert isinstance(res[0], Test3)
-    assert res[0].value == 1
-
-
-def test_unique_justseen_memoryleak():
-    unique_justseen = iteration_utilities.unique_justseen
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            return self.value == other.value
-
-    class Test2(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            raise TypeError()
-
-        def __ne__(self, other):
-            raise TypeError()
-
-    class Test3(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            raise TypeError()
-
-        def __ne__(self, other):
-            return self.value != other.value
-
-    def test():
-        list(unique_justseen([])) == []
-    assert not memory_leak(test)
-
-    def test():
-        list(unique_justseen([Test(1), Test(1), Test(2), Test(3), Test(3)]))
-    assert not memory_leak(test)
-
-    def test():
-        list(unique_justseen([Test(1), Test(-1), Test(1),
-                              Test(2), Test(-2), Test(2)]))
-    assert not memory_leak(test)
-
-    def test():
-        list(unique_justseen([Test(1), Test(-1), Test(1),
-                              Test(2), Test(-2), Test(2)],
-                             key=lambda x: abs(x.value)))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # not iterable
-            unique_justseen(Test(1))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # function call fails
-            list(unique_justseen([Test(1), Test(2), Test(3)],
-                                 key=lambda x: x + 'a'))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # objects do not support eq or ne
-            list(unique_justseen([Test2(1), Test2(2)]))
-    assert not memory_leak(test)
-
-    def test():
-        list(unique_justseen([Test3(1), Test3(1)]))
-    assert not memory_leak(test)
-
-
-def test_groupby2():
-    groupby2 = iteration_utilities.groupby2
-
-    assert groupby2([], key=lambda x: x) == {}
-    assert groupby2(['a', 'ab', 'abc'],
-                    key=operator.itemgetter(0)) == {'a': ['a', 'ab', 'abc']}
-    assert groupby2(['a', 'ba', 'ab', 'abc', 'b'],
-                    key=operator.itemgetter(0)) == {'a': ['a', 'ab', 'abc'],
-                                                    'b': ['ba', 'b']}
-    assert groupby2(['a', 'ba', 'ab', 'abc', 'b'],
-                    key=operator.itemgetter(0),
-                    keepkey=len) == {'a': [1, 2, 3], 'b': [2, 1]}
-
-    with pytest.raises(TypeError):  # not iterable
-        groupby2(1, key=len)
-
-    with pytest.raises(TypeError):  # key func fails
-        groupby2([1, 2, 3], key=lambda x: x + 'a')
-
-    with pytest.raises(TypeError):  # keepkey func fails
-        groupby2([1, 2, 3], key=lambda x: x, keepkey=lambda x: x + 'a')
-
-    with pytest.raises(TypeError):  # unhashable
-        groupby2([{'a': 10}], key=lambda x: x)
-
-
-def test_groupby2_memoryleak():
-    groupby2 = iteration_utilities.groupby2
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            return self.value == other.value
-
-        def __hash__(self):
-            return hash(self.value)
-
-    def test():
-        groupby2([], key=lambda x: x)
-    assert not memory_leak(test)
-
-    def test():
-        groupby2([Test('a'), Test('ab'), Test('abc')],
-                 key=lambda x: Test(x.value[0]))
-    assert not memory_leak(test)
-
-    def test():
-        groupby2([Test('a'), Test('ba'), Test('ab'), Test('abc'), Test('b')],
-                 key=lambda x: Test(x.value[0]))
-    assert not memory_leak(test)
-
-    def test():
-        groupby2([Test('a'), Test('ba'), Test('ab'), Test('abc'), Test('b')],
-                 key=lambda x: Test(x.value[0]),
-                 keepkey=lambda x: Test(len(x.value)))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # not iterable
-            groupby2(Test(1), key=lambda x: Test(len(x.value)))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # key func fails
-            groupby2([Test(1), Test(2), Test(3)],
-                     key=lambda x: Test(x.value + 'a'))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # keepkey func fails
-            groupby2([Test(1), Test(2), Test(3)],
-                     key=lambda x: x, keepkey=lambda x: Test(x.value + 'a'))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # unhashable
-            groupby2([{Test('a'): Test(10)}], key=lambda x: x)
-    assert not memory_leak(test)
-
-
-def test_alldistinct():
-    all_distinct = iteration_utilities.all_distinct
-
-    assert all_distinct([1, 2, 3])
-    assert not all_distinct([1, 1, 1])
-
-    assert all_distinct([{'a': 1}, {'a': 2}])
-    assert not all_distinct([{'a': 1}, {'a': 1}])
-
-    with pytest.raises(TypeError):  # iterable is not iterable
-        all_distinct(1)
-
-
-def test_alldistinct_memoryleak():
-    all_distinct = iteration_utilities.all_distinct
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            return self.value == other.value
-
-        def __hash__(self):
-            return hash(self.value)
-
-    def test():
-        all_distinct([Test(1), Test(2), Test(3)])
-    assert not memory_leak(test)
-
-    def test():
-        all_distinct([Test(1), Test(1), Test(1)])
-    assert not memory_leak(test)
-
-    def test():
-        all_distinct([{Test('a'): Test(1)}, {Test('a'): Test(2)}])
-    assert not memory_leak(test)
-
-    def test():
-        all_distinct([{Test('a'): Test(1)}, {Test('a'): Test(1)}])
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # iterable is not iterable
-            all_distinct(Test(1))
-    assert not memory_leak(test)
-
-
-def test_all_equal():
-    all_equal = iteration_utilities.all_equal
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            if type(self.value) != type(other.value):
-                raise TypeError('simulated failure.')
-            return self.value == other.value
-
-    assert all_equal([])
-    assert all_equal([1, 1, 1])
-    assert not all_equal([1, 1, 2])
-
-    with pytest.raises(TypeError):  # not iterable
-        all_equal(1)
-
-    if not iteration_utilities.PY2:
-        with pytest.raises(TypeError):  # comparison fail
-            all_equal([Test(1), Test('a')])
-
-
-def test_all_equal_memoryleak():
-    all_equal = iteration_utilities.all_equal
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            if type(self.value) != type(other.value):
-                raise TypeError('simulated failure.')
-            return self.value == other.value
-
-    def test():
-        all_equal([])
-    assert not memory_leak(test)
-
-    def test():
-        all_equal([Test(1), Test(1), Test(1)])
-    assert not memory_leak(test)
-
-    def test():
-        all_equal([Test(1), Test(1), Test(2)])
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # not iterable
-            all_equal(Test(1))
-    assert not memory_leak(test)
-
-    if not iteration_utilities.PY2:
-        def test():
-            with pytest_raises(TypeError):  # comparison fail
-                all_equal([Test(1), Test('a')])
-        assert not memory_leak(test)
-
-
-def test_split():
-    split = iteration_utilities.split
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            if type(self.value) != type(other.value):
-                raise TypeError('simulated failure.')
-            return self.value == other.value
-
-    assert list(split([], lambda x: False)) == []
-
-    assert list(split([1, 2, 3], lambda x: x == 2)) == [[1], [3]]
-    assert list(split([1, 2, 3], lambda x: x == 3)) == [[1, 2]]
-
-    # keep
-    assert list(split([1, 2, 3], lambda x: x == 2,
-                      keep=True)) == [[1], [2], [3]]
-    assert list(split([1, 2, 3], lambda x: x == 3,
-                      keep=True)) == [[1, 2], [3]]
-
-    # maxsplit
-    assert list(split([1, 2, 3, 4, 5], lambda x: x % 2 == 0,
-                      maxsplit=1)) == [[1], [3, 4, 5]]
-    assert list(split([1, 2, 3, 4, 5], lambda x: x % 2 == 0,
-                      maxsplit=2)) == [[1], [3], [5]]
-
-    # equality
-    assert list(split([1, 2, 3, 2, 5], 2,
-                      eq=True)) == [[1], [3], [5]]
-
-    # failures
-    with pytest.raises(TypeError):  # not iterable
-        split(1, lambda x: False)
-    with pytest.raises(TypeError):  # func fails
-        list(split([1, 2, 3], lambda x: x + 'a'))
-    with pytest.raises(TypeError):  # cmp fails
-        list(split([Test(1), Test(2), Test(3)], Test('a'), eq=True))
-
-
-def test_split_memoryleak():
-    split = iteration_utilities.split
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-        def __eq__(self, other):
-            if type(self.value) != type(other.value):
-                raise TypeError('simulated failure.')
-            return self.value == other.value
-
-    def test():
-        list(split([], lambda x: False)) == []
-    assert not memory_leak(test)
-
-    def test():
-        list(split([Test(1), Test(2), Test(3)], lambda x: x.value == 2))
-    assert not memory_leak(test)
-
-    def test():
-        list(split([Test(1), Test(2), Test(3)], lambda x: x.value == 3))
-    assert not memory_leak(test)
-
-    # keep
-    def test():
-        list(split([Test(1), Test(2), Test(3)], lambda x: x.value == 2,
-                   keep=True))
-    assert not memory_leak(test)
-
-    def test():
-        list(split([Test(1), Test(2), Test(3)], lambda x: x.value == 3,
-                   keep=True))
-    assert not memory_leak(test)
-
-    # maxsplit
-    def test():
-        list(split([Test(1), Test(2), Test(3), Test(4), Test(5)],
-                   lambda x: x.value % 2 == 0, maxsplit=1))
-    assert not memory_leak(test)
-
-    def test():
-        list(split([Test(1), Test(2), Test(3), Test(4), Test(5)],
-                   lambda x: x.value % 2 == 0, maxsplit=2))
-    assert not memory_leak(test)
-
-    # equality
-    def test():
-        list(split([Test(1), Test(2), Test(3), Test(2), Test(5)], Test(2),
-                   eq=True))
-    assert not memory_leak(test)
-
-    # failures
-    def test():
-        with pytest_raises(TypeError):  # not iterable
-            split(Test(1), lambda x: False)
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # func fails
-            list(split([Test(1), Test(2), Test(3)],
-                       lambda x: Test(x.value + 'a')))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # cmp fails
-            list(split([Test(1), Test(2), Test(3)], Test('a'), eq=True))
-    assert not memory_leak(test)
 
 
 @pytest.mark.xfail(iteration_utilities.PY2,
